extracted InfoMessage and errorMessage to separate module for future reuse
